ES6
The project now uses ES6 modules with a script type of module in the browser. This confirmed works in current Chrome, Firefox, and Edge. In order to support older browsers, a script type of nodemodule will need to be added to the client pages, and a new script compiled using something like requirejs or browserfy etc...

This is how I wanted things to work from the beginning for the code to be as straight forward as possible, but later bundle software, or client side module handlers should be implemented.